@@ -11,7 +11,7 @@
     "build:server": "cross-env NODE_ENV=production webpack --config build/webpack.server.config.js --progress --color --colors --hide-modules"
   },
   "keywords": [],
-  "author": "YY TANG <tlyspa@gmail.com>",
+  "author": "baotang <tlyspa@gmail.com>",
   "license": "ISC",
   "devDependencies": {
     "babel-cli": "^6.24.1",
@@ -36,15 +36,16 @@
     "sw-precache-webpack-plugin": "^0.9.1",
     "url-loader": "^0.5.8",
     "vue-loader": "^11.3.4",
-    "vue-ssr-webpack-plugin": "^2.1.0",
-    "vue-template-compiler": "^2.2.6",
+    "vue-template-compiler": "^2.3.2",
     "webpack": "^2.3.3",
     "webpack-dev-middleware": "^1.10.1",
     "webpack-hot-middleware": "^2.18.0",
-    "webpack-merge": "^4.1.0"
+    "webpack-merge": "^4.1.0",
+    "webpack-node-externals": "^1.6.0"
   },
   "dependencies": {
     "autoprefixer": "^6.7.7",
+    "axios": "^0.16.1",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "compression": "^1.6.2",
     "cross-env": "^4.0.0",
@@ -52,11 +53,11 @@
     "express": "^4.15.2",
     "normalize.css": "^6.0.0",
     "serve-favicon": "^2.4.2",
-    "vue": "^2.2.6",
-    "vue-router": "^2.4.0",
-    "vue-server-renderer": "^2.2.6",
-    "vue-style-loader": "^2.0.5",
-    "vuex": "^2.2.1",
+    "vue": "^2.3.2",
+    "vue-router": "^2.5.3",
+    "vue-server-renderer": "^2.3.2",
+    "vue-style-loader": "^3.0.1",
+    "vuex": "^2.3.1",
     "vuex-router-sync": "^4.1.2"
   }
 }
