@@ -1,34 +1,46 @@
-import 'babel-polyfill'
+// import 'babel-polyfill'
 import 'normalize.css/normalize.css'
+import './assets/less/basic.less'
 import Vue from 'vue'
 import App from './App.vue'
-import store from './store'
-import router from './router'
+import { createStore } from './store'
+import { createRouter } from './router'
 import { sync } from 'vuex-router-sync'
+// import titleMixin from './util/title'
+import filters from './filters'
 import components from './components'
 import directives from './directives'
-import filters from './filters'
-import './assets/less/basic.less'
 
-// sync the router with the vuex store.
-// this registers `store.state.route`
-sync(store, router)
+// mixin for handling title
+// Vue.mixin(titleMixin)
 
-// register global dependencies.
+
 Vue.use(components)
 Vue.use(directives)
 Vue.use(filters)
 
-// create the app instance.
-// here we inject the router and store to all child components,
-// making them available everywhere as `this.$router` and `this.$store`.
-const app = new Vue({
+// Expose a factory function that creates a fresh set of store, router,
+// app instances on each call (which is called for each SSR request)
+export function createApp () {
+  // create store and router instances
+  const store = createStore()
+  const router = createRouter()
+
+  // sync the router with the vuex store.
+  // this registers `store.state.route`
+  sync(store, router)
+
+  // create the app instance.
+  // here we inject the router, store and ssr context to all child components,
+  // making them available everywhere as `this.$router` and `this.$store`.
+  const app = new Vue({
     router,
     store,
     render: h => h(App)
-})
+  })
 
-// expose the app, the router and the store.
-// note we are not mounting the app here, since bootstrapping will be
-// different depending on whether we are in a browser or on the server.
-export { app, router, store }
+  // expose the app, the router and the store.
+  // note we are not mounting the app here, since bootstrapping will be
+  // different depending on whether we are in a browser or on the server.
+  return { app, router, store }
+}
