@@ -1,48 +1,61 @@
 const fs = require('fs')
 const path = require('path')
+const LRU = require('lru-cache')
 const express = require('express')
 const favicon = require('serve-favicon')
 const compression = require('compression')
 const resolve = file => path.resolve(__dirname, file)
+const { createBundleRenderer } = require('vue-server-renderer')
 
 const isProd = process.env.NODE_ENV === 'production'
+const useMicroCache = process.env.MICRO_CACHE !== 'false'
 const serverInfo =
-    `express/${require('express/package.json').version} ` +
-    `vue-server-renderer/${require('vue-server-renderer/package.json').version}`
+  `express/${require('express/package.json').version} ` +
+  `vue-server-renderer/${require('vue-server-renderer/package.json').version}`
 
 const app = express()
 
-let renderer
-if (isProd) {
-    // In production: create server renderer using server bundle and index HTML
-    // template from real fs.
-    // The server bundle is generated by vue-ssr-webpack-plugin.
-    const bundle = require('./dist/vue-ssr-bundle.json')
-        // src/index.template.html is processed by html-webpack-plugin to inject
-        // build assets and output as dist/index.html.
-    const template = fs.readFileSync(resolve('./dist/index.html'), 'utf-8')
-    renderer = createRenderer(bundle, template)
-} else {
-    // In development: setup the dev server with watch and hot-reload,
-    // and create a new renderer on bundle / index template update.
-    require('./build/setup-dev-server')(app, (bundle, template) => {
-        renderer = createRenderer(bundle, template)
-    })
+const template = fs.readFileSync(resolve('./src/index.template.html'), 'utf-8')
+
+function createRenderer (bundle, options) {
+  // https://github.com/vuejs/vue/blob/dev/packages/vue-server-renderer/README.md#why-use-bundlerenderer
+  return createBundleRenderer(bundle, Object.assign(options, {
+    template,
+    // for component caching
+    cache: LRU({
+      max: 1000,
+      maxAge: 1000 * 60 * 15
+    }),
+    // this is only needed when vue-server-renderer is npm-linked
+    basedir: resolve('./dist'),
+    // recommended for performance
+    runInNewContext: false
+  }))
 }
 
-function createRenderer(bundle, template) {
-    // https://github.com/vuejs/vue/blob/dev/packages/vue-server-renderer/README.md#why-use-bundlerenderer
-    return require('vue-server-renderer').createBundleRenderer(bundle, {
-        template,
-        cache: require('lru-cache')({
-            max: 1000,
-            maxAge: 1000 * 60 * 15
-        })
-    })
+let renderer
+let readyPromise
+if (isProd) {
+  // In production: create server renderer using built server bundle.
+  // The server bundle is generated by vue-ssr-webpack-plugin.
+  const bundle = require('./dist/vue-ssr-server-bundle.json')
+  // The client manifests are optional, but it allows the renderer
+  // to automatically infer preload/prefetch links and directly add <script>
+  // tags for any async chunks used during render, avoiding waterfall requests.
+  const clientManifest = require('./dist/vue-ssr-client-manifest.json')
+  renderer = createRenderer(bundle, {
+    clientManifest
+  })
+} else {
+  // In development: setup the dev server with watch and hot-reload,
+  // and create a new renderer on bundle / index template update.
+  readyPromise = require('./build/setup-dev-server')(app, (bundle, options) => {
+    renderer = createRenderer(bundle, options)
+  })
 }
 
 const serve = (path, cache) => express.static(resolve(path), {
-    maxAge: cache && isProd ? 60 * 60 * 24 * 30 : 0
+  maxAge: cache && isProd ? 60 * 60 * 24 * 30 : 0
 })
 
 app.use(compression({ threshold: 0 }))
@@ -52,34 +65,70 @@ app.use('/public', serve('./public', true))
 app.use('/manifest.json', serve('./manifest.json', true))
 app.use('/service-worker.js', serve('./dist/service-worker.js'))
 
-app.get('*', (req, res) => {
-    if (!renderer) {
-        return res.end('waiting for compilation... refresh in a moment.')
+// 1-second microcache.
+// https://www.nginx.com/blog/benefits-of-microcaching-nginx/
+const microCache = LRU({
+  max: 100,
+  maxAge: 1000
+})
+
+// since this app has no user-specific content, every page is micro-cacheable.
+// if your app involves user-specific content, you need to implement custom
+// logic to determine whether a request is cacheable based on its url and
+// headers.
+const isCacheable = req => useMicroCache
+
+function render (req, res) {
+  const s = Date.now()
+
+  res.setHeader("Content-Type", "text/html")
+  res.setHeader("Server", serverInfo)
+
+  const handleError = err => {
+    if (err && err.code === 404) {
+      res.status(404).end('404 | Page Not Found')
+    } else {
+      // Render Error Page or Redirect
+      res.status(500).end('500 | Internal Server Error')
+      console.error(`error during render : ${req.url}`)
+      console.error(err.stack)
     }
+  }
 
-    const s = Date.now()
-
-    res.setHeader('Content-Type', 'text/html')
-    res.setHeader('Server', serverInfo)
-
-    const errorHandler = err => {
-        if (err && err.code === 404) {
-            res.status(404).end('404 | Page Not Found')
-        } else {
-            // Render Error Page or Redirect
-            res.status(500).end('500 | Internal Server Error')
-            console.error(`error during render : ${req.url}`)
-            console.error(err)
-        }
+  const cacheable = isCacheable(req)
+  if (cacheable) {
+    const hit = microCache.get(req.url)
+    if (hit) {
+      if (!isProd) {
+        console.log(`cache hit!`)
+      }
+      return res.end(hit)
     }
+  }
 
-    renderer.renderToStream({ url: req.url })
-        .on('error', errorHandler)
-        .on('end', () => console.log(`whole request: ${Date.now() - s}ms`))
-        .pipe(res)
+  const context = {
+    title: 'Vue HN 2.0', // default title
+    url: req.url
+  }
+  renderer.renderToString(context, (err, html) => {
+    if (err) {
+      return handleError(err)
+    }
+    res.end(html)
+    if (cacheable) {
+      microCache.set(req.url, html)
+    }
+    if (!isProd) {
+      console.log(`whole request: ${Date.now() - s}ms`)
+    }
+  })
+}
+
+app.get('*', isProd ? render : (req, res) => {
+  readyPromise.then(() => render(req, res))
 })
 
 const port = process.env.PORT || 8080
 app.listen(port, () => {
-    console.log(`server started at localhost:${port}`)
+  console.log(`server started at localhost:${port}`)
 })
